Text scan function reads the ASCII input of data and changes it to MATLAB.  A few edits need to be made for different data sets. Like the filename needs to be updated in the function, the number of headlines to be ignored before reading the actual data, and lastly the different data types where a double, or a float.
Text scan function reads the ASCII input of data and changes it to MATLAB.  A few edits need to be made for different data sets. Like the filename needs to be updated in the function, the number of headlines to be ignored before reading the actual data, and lastly the different data types where a double, or a float.